requirements-test.txt
-r requirements.txt
coverage
coveralls
mock>=1.0.1
nose>=1.3.0
django-nose>=1.2

# Additional test requirements go here
WebTest==2.0.14
